% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonstat_methods.R
\name{as.json}
\alias{as.json}
\title{Convert to jsonlite json object}
\usage{
as.json(x, ...)
}
\arguments{
\item{x}{an object to coerce to a json object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Convert to jsonlite json object
}
\details{
Currently only methods for \code{jsonstat} objects are implemented.
}
