% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonstat_methods.R
\name{as.vector.jsonstat}
\alias{as.vector.jsonstat}
\title{Convert json stat object to numeric vector}
\usage{
as.vector.jsonstat(x, mode = "any")
}
\description{
Convert json stat object to numeric vector
}
\details{
This converts a json stat object to a vector. It essentially just
returns the value slot of a json stat object.
}
\seealso{
as.vector
}

