% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonstat_methods.R
\name{as.vector.jsonstat_dataset}
\alias{as.vector.jsonstat_dataset}
\title{Convert json stat object to numeric vector}
\usage{
as.vector.jsonstat_dataset(x, mode = "any")
}
\arguments{
\item{x}{a \code{jsonstat_dataset} object}

\item{mode}{character string naming an atomic mode or "\code{list}" or
"\code{expression}" or (except for \code{vector}) "\code{any}".}
}
\description{
Convert json stat object to numeric vector
}
\details{
This converts a json stat object to a vector. It essentially just
returns the value slot of a json stat object.
}
\seealso{
\code{\link{as.vector}}
}

