% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjstat.R
\name{fromJSONstat}
\alias{fromJSONstat}
\title{Convert JSON-stat format to data frame(s)}
\usage{
fromJSONstat(x, naming = "label", use_factors = FALSE)
}
\arguments{
\item{x}{the JSON-stat format in characters (or path or URL)}

\item{naming}{whether to use (longer) \code{label}s or (shorter) \code{id}s}

\item{use_factors}{whether dimension categories should be factors or
character objects}
}
\value{
For responses with class \code{dataset} or \code{bundle}: A list of
  one or more data frames. For responses with class \code{collection}: A list
  of one list of one or more lists or data frames.
}
\description{
This function takes characters of a JSON-stat format response (or path
or URL to such a response) and returns a list of data frames with
columns for each dimension and one \code{value} column.
}
\examples{
\dontrun{
oecd.canada.url <- "http://json-stat.org/samples/oecd-canada.json"
results <- fromJSONstat(oecd.canada.url)
names(results)
}
}

