% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonstat_accessors.R
\name{id}
\alias{id}
\alias{id<-}
\title{\code{id} accessors of \code{jsonstat} objects}
\usage{
id(x)

id(x) <- value
}
\arguments{
\item{x}{a \code{jsonstat_dataset} object}

\item{value}{a character vector of up to the same length as \code{id(x)}.}
}
\description{
Access and change \code{id}s of jsonstat objects.
}
