% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonstat.R
\name{jsonstat}
\alias{jsonstat}
\title{Create a JSON-stat object}
\usage{
jsonstat(x)
}
\arguments{
\item{x}{a JSON-stat string, URL or file}
}
\value{
a JSON-stat object with subclass dataset, dimension or collection
}
\description{
Create a JSON-stat object
}
\examples{
oecd_jsonstat <- jsonstat("http://json-stat.org/samples/oecd.json")

}

