% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjstat.R
\docType{package}
\name{rjstat}
\alias{rjstat}
\alias{rjstat-package}
\title{Read and write JSON-stat data sets}
\description{
\href{http://json-stat.org}{JSON-stat} is a JSON format for data
dissemination. The \pkg{rjstat} package converts data frames to and from this
format. The extensive metadata features of JSON-stat are not supported.
}

