% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonstat_accessors.R
\name{status}
\alias{status}
\alias{status<-}
\title{\code{status} accessors of \code{jsonstat} objects}
\usage{
status(x)

status(x) <- value
}
\arguments{
\item{x}{a \code{jsonstat_dataset} object}

\item{value}{a character vector of up to the same length as \code{as.vector(x)} or a named list with index as vector names.}
}
\description{
Access and change \code{status}s of jsonstat objects.
}
