% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonstat_methods.R
\name{unbox}
\alias{unbox}
\title{Unbox generic function}
\usage{
unbox(x)
}
\arguments{
\item{x}{an object to unbox}
}
\description{
Unbox generic function
}
\details{
This is a generic function to unbox more general objects,
such as \code{jsonstat} objects.
See \code{\link[jsonlite]{unbox}} for mor general information.
}
\seealso{
\code{\link[jsonlite]{unbox}}
}

