% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonstat.R
\name{validate_jsonstat}
\alias{validate_jsonstat}
\title{Function to validate json stat}
\usage{
validate_jsonstat(x)
}
\description{
Now this is just a simple light-weight validation.
Hopefully this can be complemented with a real json stat
schema validator.
}
\keyword{internal}

